+++ server.js
@@ -43,6 +43,7 @@ const coupon = require('./routes/coupon')
 const basket = require('./routes/basket')
 const order = require('./routes/order')
 const verify = require('./routes/verify')
+const recycles = require('./routes/recycles')
 const b2bOrder = require('./routes/b2bOrder')
 const showProductReviews = require('./routes/showProductReviews')
 const createProductReviews = require('./routes/createProductReviews')
@@ -185,10 +186,10 @@ app.get('/api/Complaints', insecurity.isAuthorized())
 app.post('/api/Complaints', insecurity.isAuthorized())
 app.use('/api/Complaints/:id', insecurity.denyAll())
 /* Recycles: POST and GET allowed when logged in only */
-app.get('/api/Recycles', verify.blockRecycleItems())
+app.get('/api/Recycles', recycles.blockRecycleItems())
 app.post('/api/Recycles', insecurity.isAuthorized())
 /* Challenge evaluation before epilogue takes over */
-app.get('/api/Recycles/:id', verify.sequelizeVulnerabilityChallenge())
+app.get('/api/Recycles/:id', recycles.sequelizeVulnerabilityChallenge())
 app.put('/api/Recycles/:id', insecurity.denyAll())
 app.delete('/api/Recycles/:id', insecurity.denyAll())
 /* SecurityQuestions: Only GET list of questions allowed. */
